+++ datadog_checks_base/tests/base/checks/test_fips_mode.py
@@ -3,16 +3,21 @@
 # (C) Datadog, Inc. 2024-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
+import os
 import subprocess
 import sys
-from pathlib import Path
 from typing import Any  # noqa: F401
 
 import pytest
+import requests
 
 from datadog_checks.base import AgentCheck
+from datadog_checks.dev import docker_run, get_here
+from datadog_checks.dev.conditions import WaitForPortListening
 
-PATH_TO_EMBEDDED = str(Path(__file__).resolve().parent.parent.parent / "fixtures" / "fips" / "embedded")
+HERE = get_here()
+EMBEDDED = os.path.join(HERE, '..', '..', 'fixtures', 'fips', 'embedded')
+FIPS_SERVER_PORT = 443
 
 
 @pytest.fixture(scope="session", autouse=True)
@@ -21,9 +26,9 @@ def create_fipsmodule_config():
         "openssl",
         "fipsinstall",
         "-module",
-        f'{PATH_TO_EMBEDDED}/lib/ossl-modules/fips.so',
+        f'{EMBEDDED}/lib/ossl-modules/fips.so',
         "-out",
-        f'{PATH_TO_EMBEDDED}/ssl/fipsmodule.cnf',
+        f'{EMBEDDED}/ssl/fipsmodule.cnf',
         "-provider_name",
         "fips",
     ]
@@ -55,7 +60,7 @@ def test_md5_before_fips():
 def test_md5_after_fips(clean_environment):
     import ssl
 
-    AgentCheck().enable_openssl_fips(path_to_embedded=PATH_TO_EMBEDDED)
+    AgentCheck().enable_openssl_fips(path_to_embedded=EMBEDDED)
     with pytest.raises(ssl.SSLError, match='No cipher can be selected.'):
         ctx = ssl.SSLContext(ssl.PROTOCOL_TLS_CLIENT)
         ctx.set_ciphers("MD5")
@@ -73,6 +78,44 @@ def test_cryptography_md5_fips():
     from cryptography.exceptions import InternalError
     from cryptography.hazmat.primitives import hashes
 
-    AgentCheck().enable_cryptography_fips(path_to_embedded=PATH_TO_EMBEDDED)
+    AgentCheck().enable_cryptography_fips(path_to_embedded=EMBEDDED)
     with pytest.raises(InternalError, match='Unknown OpenSSL error.'):
         hashes.Hash(hashes.MD5())
+
+
+@pytest.fixture(scope="session")
+def non_fips_server():
+    conditions = [WaitForPortListening("localhost", 443)]
+    with docker_run(os.path.join(HERE, 'docker', 'fips_compose.yaml'), conditions=conditions):
+        yield
+
+
+@pytest.mark.skipif(not sys.platform == "linux", reason="only testing on Linux")
+def test_connection_before_fips():
+    """
+    Test connection to the FIPS server before enabling FIPS mode.
+    """
+    url = f"https://localhost:{FIPS_SERVER_PORT}"
+    try:
+        response = requests.get(url, verify=False, timeout=5)
+        assert response.status_code == 200
+    except requests.exceptions.SSLError as e:
+        pytest.fail(f"Connection failed due to SSL error: {e}")
+
+
+@pytest.mark.skipif(not sys.platform == "linux", reason="only testing on Linux")
+def test_connection_after_fips():
+    """
+    Test connection to the FIPS server after enabling FIPS mode.
+    """
+    os.environ["OPENSSL_CONF"] = os.path.join(EMBEDDED, 'ssl', 'openssl.cnf')
+    os.environ["OPENSSL_MODULES"] = os.path.join(EMBEDDED, 'lib', 'ossl-modules')
+    os.environ["GOFIPS"] = "1"
+    AgentCheck()
+
+    url = f"https://localhost:{FIPS_SERVER_PORT}"
+    try:
+        response = requests.get(url, verify=False, timeout=5)
+        assert response.status_code != 200
+    except requests.exceptions.SSLError as e:
+        pytest.fail(f"Connection failed due to SSL error: {e}")
